stdlib: Fix a bug concerning exception class 'throw'
Since the introduction of the stack trace variable, the Erlang Pretty
Printer has left out the exception class 'throw' even when the stack
trace variable cannot be left out, which is not correct Erlang code.
The fix is to always include the exception class 'throw'.

An alternative is to exclude 'throw' exactly when the stack trace
variable is '_'.