Report stacktrace in non-transactions activity function errors.
If an activity function fails with an error, the stacktrace is
dropped in non_transaction function try/catch clause.
This makes debuging of errors inside a transaction really hard.

Transaction activities handle errors and exits differently, non-transaction
activities should do the same.

Adding the stacktrace to the exit reason when translating errors to exits.