erts: Fix so that non-oneshot pollset on OpenBSD works
The poll code for kevent on OpenBSD and NetBSD had not been fixed
to work properly when disabling events in the scheduler poll-set.
Because they use ONESHOT the delete operation was built into the
trigger and thus a lot of code was not needed when doing oneshot,
however it is needed when doing non-oneshot.