Make the `cpu_sup_SUITE:util_values` test case more reliable
Previously, the `util_values` test case in the test suite
`cpu_sup_SUITE` tested the `cpu_sup:util()` by checking if a process
that spins in a loop cause the CPU utilization measured with
`cpu_sup:util()` to increase. This was unreliable on test machines
that ran other tasks at the same time. This commit tries to make the
test case more reliable by skipping the test case if it is detected
that the system is doing other work that use a lot of CPU time and
starting 100 spinning processes instead of only 1.