Don't treat reception of NETCONF > 1.0 hello as error
Module ct_netconfc appears to have followed the introduction of NETCONF
1.1 in RFC 6241, but assumes NETCONF 1.0 (RFC 4741). A server supporting
the 1.1 can send the corresponding base capability in hello, which
caused ct_netconfc to close the connection. Now simply ignore any
version but 1.0, and fail only if the server doesn't advertise 1.0.

Actual support for 1.1 is still missing: being able to agree on this
version with the peer and receive messages that are chunked as specified
in RFC 6242.