Add ct_netconfc support for NETCONF 1.1
The 1.1 base capability can now be passed to hello/3, and the version
selected for the session is as required by RFC 6241, 8.1 Capabilities
Exchange:

   Both NETCONF peers MUST verify that the other peer has advertised a
   common protocol version.  When comparing protocol version capability
   URIs, only the base part is used, in the event any parameters are
   encoded at the end of the URI string.  If no protocol version
   capability in common is found, the NETCONF peer MUST NOT continue the
   session.  If more than one protocol version URI in common is present,
   then the highest numbered (most recent) protocol version MUST be used
   by both peers.

Negotiation of 1.1 (or any version greater than 1.0) results in message
being chunked as required by RFC 6242, instead of the 1.0 framing
(]]>]]>) specified by RFC 4742.

For backwards compatibility, the 1.0 base capability is added to any
capabilities specified by the user only if no base capability is
specified. That is, no user-specified base capability results in 1.0
being added, otherwise it's up to the user to specify all base
capabilities.