Use a more reasonable scratch register in the compiler
This uses a X register larger than any used normally in the function
for the scratch register when shuffling registers. In many cases,
there's probably an even lower register that could be used, but this
is a simple measure that avoids using the 1022 register. The high
register was both alarming when analysing the produced code and might
possibly lower runtime performance by significantly inflating the number
of used registers.

Additonally simplify the algorithm for finding scratch register when
shuffling call arguments.