Let capabilities be set for all session establishment
Commit e69da353 added the possibility of passing capabilities to
hello/3, but didn't allow them to be passed into open/2. Not strictly
necessary, but simpler than having to call only_open/2 followed by
hello/3 if all that's wanted is to add capabilties to the hello message.

Commit 63d63920 then added session/3 for multiple channels per
connection, but no way of passing capabilities or omitting hello, the
former now being a problem if the user wants to send the NETCONF 1.1
base capability.

Give the user control over capabilities by allowing them to be passed to
session/3, and to open/2 for consistency, but don't do anything about
avoiding hello in session/3.

That capabilities have to be specified as a list of strings makes little
sense for the user, but is because commit e69da353 passed what
xmerl:export_simple/3 wanted. Also leave this as is for now.

Don't document server capabilities since these depend on the NETCONF
version.

The history isn't clear, but an option to delay hello by some specified
amount of time (or indefinitely) may have been a simpler alternative to
what only_open/hello are meant to achieve.