Always fallback to source when we can't parse AST
This commit fixes a bug in Dialyzer where it would assume
all Erlang AST can be converted back to valid Erlang code.
A case where this was is no true was already documented
in the source code, in relation to opaque types, but this
can also be true when it comes to other BEAM languages.