Improve optimization of redundant tests
The `beam_ssa_dead` pass is supposed to eliminate tests that are
determined to be redundant based on the outcome of a previous test.

For example, in the following example that repeats a guard test,
the second clause can never be executed:

    foo(A) when A >= 42 -> one;
    foo(A) when A >= 42 -> two;
    foo(_) -> three.

`beam_ssa_dead` should have eliminated the second clause, but
didn't:

      {test,is_ge,{f,5},[{x,0},{integer,42}]}.
      {move,{atom,one},{x,0}}.
      return.
    {label,5}.
      {test,is_ge,{f,6},[{x,0},{integer,42}]}.
      {move,{atom,two},{x,0}}.
      return.
    {label,6}.
      {move,{atom,three},{x,0}}.
      return.

Correct the optimization of four different combinations of relational
operations that were too conservate. To ensure the correctness of the
optimization, also add an exahaustive test of all combinations of
relational operations with one variable and one literal. (Also remove
the weak and now redundant coverage tests in `beam_ssa_SUITE`.) With
this correction, the following code will be generated for the example:

      {test,is_ge,{f,5},[{x,0},{integer,42}]}.
      {move,{atom,one},{x,0}}.
      return.
    {label,5}.
      {move,{atom,three},{x,0}}.
      return.

Thanks to Dániel Szoboszlay (@dszoboszlay), whose talk at Code BEAM
STO 2019 made me aware of this missed opportunity for optimization.