Add script to drop into development container
Start an interactive session inside a container based on the development
Docker image. Ensure a fixed hostname is set for the container to avoid
having to reconfigure tests etc.

Mount the local repo into the container buildroot to allow working with
local tools on the host machine while running builds and test inside the
container.