[snmp|test] Improve the test case start
The function used by "all" the agent test cases to
actually run the operations have been improved.
There was previously very little monitoring of the
result. Have added some (minor) checks, both before
trying to running the test case, and "during".
Such as, is the node we attempt to use actually alive.
Then, when we spawn the test case runner process on
the (remote) node, make it report back before trying
to run the actuall test case (so we know that the spawn
worked.
Also added a monitor of the process, so that we will
detect fatal errors.

OTP-15764