Optimise stacktrace construction using literals
Pre-allocate all file names for each module as literals when
the module is loaded. This saves on copying the file names at
runtime when the stacktrace is constructed as well as copying
when sending to other processes on failure - this might
especially benefit supervisors that can often receive a lot of
stacktraces in EXIT messages.