[snmp] Add verbosity printouts to the set handling
Add verbosity printouts after a set operation. Both
for a successful and a failed set.

On one of our test hosts (a VM running an old gento),
it looked like an ets or dets insert hangs. But since
we only have a printout directly before, we can't
tell for sure what happens.

OTP-15764