Attempt at fixing AIX while keeping API compatibility
erl_interface has a function in a struct, and AIX headers, when you
use C instead of C++, pollute the global namespace with definitions
that override socket functions, and end up clobbering names of
things Erlang/OTP uses. The new socket interface of R22 was one of
the victims.  I had previously tied to fix it for erl_interface,
but the change broke API public surface. What this does is replace
the name with a define, that defines to the same name on non-AIX,
but uses a "mangled" name on AIX instead. This means internal uses
of this function in the struct need to use the preprocessor define
instead of the "normal" name, and likewise external users too,
but those can be dealt with one-at-a-time on AIX instead of
breaking everyone else.