erts: Cleanup start of all erts threads to ABORT when failing
Before this change erts used to crash dump and then abort, but a
crash dump is not really usefull at this point and it caused all
sort of lock problems when crash dumping that early in the system
boot, so now it is changed to only dump core instead.

Also in the process I cleaned up some the code so that it does
not to a lot of things that are not needed.