[megaco|test] Timer test suite improvements
We did not check the actual result of the timer cancel
operation before continuing, which seems to cause
problems on some platforms (where the timer may actually
expire even though we have 100 ms to cancel).
So now we check the cancel result, and if we did not
succeed, we instead skip the test case.