[megaco|test] Fixing the examples test suite
The simple test case was faulty in that the simple example
(which it is testing) require that the MGC and MG run on different
erlang nodes. But the test case did not do this. Instead, it ran
both the MGC and the MG on the test server node, which caused
the test case to fail 1 time of 10.