[megaco] Add megaco_user behaviour module
Add a "proper" behaviour module for the megaco_user behaviour.
This behaviour is not a proper behaviour since its possible to
configure megaco to add additional arguments to *all* of the
functions in the behaviour (with the user_args config option).
So, this behaviour only reflects the *default config* (whith
the user_args set to [], that is no extra args).

OTP-15882