[snmp|manager|test] Problems starting (slave) nodes
When we fail to start a (slave) node with reason timeout,
its possible that it actually succeeded, which will cause
the following test case(s) init to fail (with 'already started',
since it which also try to start this node).
But since we don't know if we get this because for this reason
or because the previous test case failed to clean up after
itself, we cannot assume that the state of the node is ok
(and therefor use it). So, we attempt to stop it, and try
start it again (*one* time).

This has been observed on one (slooow) VM host.