The map() type is an alias; not a predefined type
This pull request corrects an historical accident.
At some point in time, the set of Erlang terms was extended to include
maps, and at that time the `map()' type was also introduced to the
language as an easy, albeit quite coarse, way to declare map terms.
At some later point in time, the type language was extended with the
ability to declare _type associations_ in maps and also with a notation
to declare mandatory (:=) and optional (=>) type associations.

At that point in time, it should have been realized that the `map()'
type is nothing more than an alias (for `#{any() => any()}') instead
of being a _predefined_ type.  (The situation is analogous to e.g.
the `binary()' type being just an alias for `<<_:_*8>>'.)

This pull request corrects the documentation (the reference manual).