Authenticate with an RSA key added to SSH Agent
Temporarily replace some existing functions to use ssh-agent and get it
to create a signature with an added RSA key.

We will have to generalize this behavior and re-integrate the old
functionality for cases where no agent should be used (default).

Additionally, different signature formats depending on the key algorithm
need to be handled (currently only RSA signatures are decoded properly)
and we need to look into concurrent agent requests.