Implement "ssh_client_key_api" with agent support
Add a module implementing the "ssh_client_key_api" behaviour, directing
user key queries to an SSH agent and delegating calls to host key
callbacks to the default "ssh_file" module.

With this approach, the "ssh_client_key_api:user_key/2" signature needs
to be extended to allow returning "partial keys" from an agent that do
not contain the private key part.

This is more of a short-term solution trying to blend in with the
existing APIs. However, these have a few short-comings that should
probably addressed in the future, e.g.

- Allow for multiple key candidates per algorithm (see RFC 4252, 5.1).
- Potentially replace the "ssh_client_key_api" behaviour to avoid
  "bleeding" of "ssh_file"- vs. "ssh_file_with_agent"-specific handling
  into the "ssh_auth" module. Currently, "ssh_auth" needs to
  differentiate between keys returned from files and keys from an agent.
  The new behaviour should possibly cover "signing" to achieve this, but
  these are just some early ideas.