[snmp|agent|test] Add another "skip point"
A test run where the node died after a series of "busy dist port"
events and then the node died which caused the process that ran
the test case to crash with reason = noconnection. And since this
was a "permanent" node, the remaining test cases then crashed.
Again, this has nothing to do with snmp or its test suites.