beam_ssa_type: Skip impossible branches
Whenever there's a type conflict during type inference we know
that the branch will not be taken. Previously we'd go down the
branch with garbage type information which would crash in some
cases.

There's no test case for the crashes because the ones I've found
require union types to happen, and we already have good coverage
once they're in place.