[megaco|test] Misc improvements to the trans test (sub-) suite
The actual "improvement" has to do with io. That is, making
sure we get any. A test case failed (on a OpenIndiana VM),
which basically only pointed to which line in the test code.
But it was impossible to figure out from that what actually
had gone wrong.

Also removed the 'export all' compiler flag, which caused a
whole bunch of unused function to become visible. Commented
out those...