[megaco|test] Improve handling of skip
The load test suite did not properly handle attempts to
skip test cases. If we failed to start one of the slave
nodes for whatever reason (in this case we got a timeout),
this should not result in a failed test case as this is
out of the control of the test case. As it was, we *did*
try to issue a skip, but the "load controller" did not
detect it. It instead transformed it into a regular error.