[megaco|test] Failure to start config should result in skip
It seems that the megaco_config process did not start (for
an unknown reason) and that resulted in a failed test case.
But starting the config process was not what the test case
was for, so it should have resulted in a skip.
Also, since we did not check the result of the config start,
we don't know why it failed.
So, move the start and stop of the megaco_config process
to the [init|end]_per_testcase function(s). If we fail to
start, then we skip (with a useful reason).