[megaco|test] Message in "wrong" order cause test case fail
It looked like two messages was received in the "wrong" order,
cuasing the request_and_reply_plain test case tro fail (on
a windows machine).
On closer inspection, the reason was that an "unconditional"
receive was made when waiting for a (disconnect) callback.
But instead we received a secondary message. So, remove
the unconditional receive and instead introduce a timeout
(10 seconds).