[snmp|manager|test] Fixed notify test case
The (second) notify test case had a faulty "condition" test
(which skipped the case in all cases except when it was
supposed to be skipped :)
Also, the test start and stop of the manager during the
test case has been adjusted. We try to keep track of how
long time each start and stop takes. And also, if they takes
to long, we skip the case.