[esock] Renaming internal functions to avoid name clashes
The naming "convention" for "top" functions of the esock
nif module was: The API functions where called nif_<foo>
This function did basic checks and arg extraction and then
called the a function called n<foo>, which was the one doing
the actual work.
Unfortunately, och some platforms (AIX) these second names,
n<foo>, clashed with system functions.
So, in order to avoid this, the naming of second function
has been changed to esock_<foo>.

OTP-15955