Avoid extremely long compilation times for huge functions
Compiling this example takes less than a second for OTP 21:

    -define(B, {?A,?A,?A,?A,?A}).
    -define(C, {?B,?B,?B,?B,?B}).
    -define(D, {?C,?C,?C,?C,?C}).
    -define(E, {?D,?D,?D}).

    f() -> ?E = foo:bar().

The compilation time for OTP 22 is about 10 seconds. Most of the
time is spent in `beam_ssa_dead`.

This commit introduces several optimizations to bring the compilation
time down to about a second.

The most important of those optimizations is limiting the effort spent
searching forward for a joining point for the success and failure
labels for a two-way branch. This change is helped by the change of
representation of variable sets from `ordsets` to `cerl_sets`.

https://bugs.erlang.org/browse/ERL-1014