Optimize deallocate_return instructions
Eliminating the CP register and putting continuation
pointers directly on the stack made the deallocate_return
instruction slower.

Try to mitigate this slow down by specializing deallocate_return
for small stack. For the move_deallocate_return instruction,
reorder instructions to make it possible to execute the read
instructions in parallel.