Optimize the move_call instructions
The `move_call` instructions are combinations of a `move` instruction
and a `call` instruction. As currently implemented, the `move` part of
the instruction is executed in its entirety before the `call` part is
even started. The reason is that the C compiler cannot see that it
would be safe to read the target address of the call before writing to
the move destination.

Rewrite the instructions to explicitly read both the source for the
move and the target address for the call before writing the
destination of the move.

Micro-benchmarks show a small but consistent speed-up after this
change.