Update test suite
Add record with exact and associated map types to type_specs.erl. The
record forces the erl_parse format for map types to be retained.

Add test wrapped_subtrees to the test suite. It ensures that each
erl_parse node it encounters can be wrapped and accessed with
erl_syntax:subtrees/1 without error.