Fix filelib:wildcard/1,2 for patterns containing ".." and/or "@"
`..` was broken and only worked when it was used in the beginning
of the pattern before any wildcard characters. For example:

    1> filelib:wildcard("erts/..").
    ["erts/.."]

Using `..` preceded by wildcard characters would not work:

    1> filelib:wildcard("*/..").
    []

`@` is not a wildcard character but is used internally in `filelib` as
an escape character and was not handled as other literal
characters. That could lead to performance degradation as it disabled
an optimization of the matching of the literal prefix of a pattern. It
would also cause the following example to fail:

    1> filelib:wildcard("@/..").
    []

This commit corrects the handling `..` and also makes sure that the
use of `@` in a pattern does not degrade performance.

https://bugs.erlang.org/browse/ERL-1029