stdlib: Fix ets_SUITE:tabfile_ext4
on machines with long host names.

Move forward the file location interval to corrupt.

An ETS table file corrupted in the header may still be readable
with table content unchanged and that will cause tabfile_ext4 to fail.
The header length increase with long host names and added table options.