Make http_uri:encode/1 percent encoding conform to RFC 3986
Previously, only characters explicitly listed as reserved were being
escaped, while others outside the ASCII range, such as 'ö' were passed
through unchanged. Also, the reserved list was wrong, both missing some
characters (e.g. '!' and '*') and listing some that are not actually
reserved (e.g. '<' and '>').

The correct behaviour is to always escape all characters that are not in the
unreserved set, i.e., A-Z, a-z, 0-9, and '-', '_', '.', '~'.