compiler: Honor stack frames when calling exit BIFs
Since exit BIFs were known not to return nor touch the stack, we
allowed them to be called regardless of how the stack looked. This
was a pretty small improvement which turned out to be a annoying
hindrance to the new BIF tracing, as we'd overwrite the return
address if we turned the specialized BIF call to an ordinary call.

This commit removes that optimization. Modules compiled before OTP
23 will still work, but exit BIFs that are traced will produce
incorrect stack traces.